safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.4_String_Literals/S7.8.4_A1.2_T2-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unexpected EOF
S7.8.4_A1.2_T2

PASS Expected parsing failure

TEST COMPLETE
